refactor: move all modals into a single folder
