+++ apps/react-ui/client/src/components/Modals/index.ts
@@ -1,0 +1,5 @@
+import MAIVEInfoModal from "./MAIVEInfoModal";
+import ParametersHelpModal from "./ParametersHelpModal";
+import RunInfoModal from "./RunInfoModal";
+
+export { MAIVEInfoModal, RunInfoModal, ParametersHelpModal };
